requirements.txt
google-genai
elevenlabs
python-dotenv
sounddevice
soundfile
numpy
speechrecognition
scipy
